Refactor CLI into proper subpackage, introduce more organization
Moves things into `cli`, a top-level package in `job_queue`.
Subcommands are refactored into one file each in src/cli/commands,
and register a subparser each in `cli/__init__.py`.

Some behavior quirks and current shortcomings are documented in comments,
and will be addressed in subsequent commits.
